requirements.dev.txt
model-bakery==1.4.0
pytest-django==4.5.2
pytest-mock==3.6.1
validators==0.19.0
